+++ src/components/post/ImageUploader.module.scss
@@ -1,4 +1,5 @@
 .uploader-wrapper {
+
     display: flex;
     flex-direction: column;
     justify-content: center;
