+++ src/components/post/PostAdd.module.scss
@@ -1,0 +1,45 @@
+@keyframes smoothAppear {
+  from {
+    opacity: 0;
+    transform: translateY(-5%);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
+
+.addBoard-wrapper {
+  opacity: 0;
+  animation: smoothAppear 1s forwards;
+  animation-delay: 0.5s;
+  font-family: 'Noto Sans KR', sans-serif;
+  .addBoard-header{
+    text-align: center;
+    font-size: 32px;
+    font-weight: bold;
+    margin: 20px 0;
+  }
+  .submitButton{
+    padding: 20px 0;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    button{
+
+    }
+    .disable-button{
+      font-size: 1.1rem;
+      cursor: not-allowed;
+    }
+    .success-button{
+      font-size: 1.1rem;
+    }
+  }
+  .addBoard-body {
+    display: flex;
+    margin: 20px 0;
+    justify-content: center;
+    flex-wrap: wrap;
+  }
+}
