+++ src/layouts/PostLayout.tsx
@@ -1,4 +1,3 @@
-import Footer from 'components/UI/Footer';
 import Header from 'components/UI/Header';
 import { Outlet } from 'react-router-dom';
