+++ src/pages/PostPage.module.scss
@@ -1,0 +1,129 @@
+// styles.scss
+
+// 전체 페이지 스타일
+$body-background-color: #d9e0d4;
+$body-hover-background-color: #b9bfb5;
+$form-background-color: #fff;
+$shadow-color: rgba(0, 0, 0, 0.1);
+
+.container {
+  background-color: $body-background-color;
+  padding: 20px;
+  border-radius: 5px;
+  width: 100vw;
+  height: 100vh;
+  box-shadow: 0 2px 5px $shadow-color;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  overflow: hidden;
+}
+
+.wrap {
+  width: 50vw;
+  height: 80%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  background-color: #fff;
+  padding: 2vh;
+  box-shadow: 3px 3px 2px 2px rgba(0, 0, 0, 0.1);
+}
+
+// 폼 요소 스타일
+.postForm {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-content: center;
+  flex-direction: column;
+}
+.label {
+  display: flex;
+  justify-content: space-around;
+  align-content: center;
+  flex-direction: row;
+  width: 100%;
+}
+.leftBox {
+  width: 20%;
+  font-size: 0.8em;
+  font-weight: 600;
+  display: flex;
+  margin-top: 1vh;
+  // justify-content: center;
+  // align-content: center;
+}
+.rightBox {
+  width: 80%;
+}
+// input박스
+.formTitle,
+.formContent,
+.formWriter,
+.formPWD {
+  width: 90%;
+  border: 1px solid #ccc;
+  height: 70%;
+}
+.formTitle {
+  height: 80%;
+}
+.formContent {
+  height: 90%;
+}
+.formCategory {
+  width: 90%;
+  height: 70%;
+  padding: 0.5%;
+}
+.form_category {
+  display: block;
+  margin-bottom: 15px;
+  height: 8vh;
+  width: 100%;
+  padding: 10px;
+  border: 1px solid #814141;
+  border-radius: 5px;
+}
+
+// form 별 높이
+.writer {
+  height: 10%;
+}
+.category {
+  height: 10%;
+}
+.title {
+  height: 10%;
+}
+.content {
+  height: 40%;
+}
+.pwd {
+  height: 10%;
+}
+// 폼 제목 스타일
+h2 {
+  font-size: 24px;
+  margin-bottom: 20px;
+}
+
+.form__submit {
+  width: 90%;
+  height: 5vh;
+  text-align: center;
+  border-radius: 50px;
+  background-color: $body-background-color;
+  border: none;
+  // margin-bottom: 5vh;
+  margin: 2vh auto;
+  letter-spacing: 1cap;
+  box-shadow: 1px 1px 2px 0 rgba(0, 0, 0, 0.1);
+}
+.form__submit:hover {
+  background-color: $body-hover-background-color;
+}
