+++ src/pages/PostViewPage.tsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import axios from 'axios';
+import Header from 'components/UI/Header';
 import styles from './PostViewPage.module.scss';
 import BoardList from '../components/board/BoardList';
